rename `fpy2.number.float` to `fpy2.number.number`
