+++ fpy2/number/mpb.py
@@ -10,7 +10,7 @@ from typing import Optional
 from ..utils import default_repr
 
 from .context import SizedContext
-from .float import Float
+from .number import Float
 from .mps import MPSContext
 from .real import RealFloat
 from .round import RoundingMode, RoundingDirection
